requirements.txt
langchain_community==0.3.24
python-dotenv==1.1.0
streamlit==1.45.1
pandas
sentence_transformers
sqlalchemy
